+++ src/relstorage/adapters/mysql/drivers/_mysqldb_gevent.py
@@ -24,10 +24,11 @@ wait = socket.wait # pylint:disable=no-member
 
 # pylint:disable=wrong-import-position,no-name-in-module,import-error
 from MySQLdb.connections import Connection as BaseConnection
-from MySQLdb.cursors import SSCursor as BaseCursor
+from MySQLdb.cursors import SSCursor
 
+from . import IterateFetchmanyMixin
 
-class Cursor(BaseCursor):
+class Cursor(IterateFetchmanyMixin, SSCursor):
     # Internally this calls mysql_use_result(). The source
     # code for that function has this comment: "There
     # shouldn't be much processing per row because mysql
@@ -82,15 +83,6 @@ class Cursor(BaseCursor):
                 break
             self.sleep()
         return result
-
-    def __iter__(self):
-        fetch = self.fetchmany
-        batch = fetch()
-        while batch:
-            for row in batch:
-                yield row
-            self.sleep()
-            batch = fetch()
 
     def enter_critical_phase_until_transaction_end(self):
         # May make multiple enters.