+++ src/relstorage/tests/__init__.py
@@ -369,7 +369,7 @@ class LockException(Exception):
     pass
 
 class MockDriver(object):
-
+    cursor_arraysize = 64
     disconnected_exceptions = (DisconnectedException,)
     close_exceptions = (CloseException,)
     lock_exceptions = (LockException,)