+++ CHANGES.rst
@@ -37,6 +37,15 @@
   ``--prepack`` and ``--use-prepack-state`` to the RelStorage, even
   when it has been wrapped in a ``zc.zlibstorage``.
 
+- Reduce the amount of memory required to pack a RelStorage through
+  more careful datastructure choices. On CPython 3, the peak
+  memory usage of the prepack phase can be up to 9 times less. On
+  CPython 2, pre-packing a 30MM row storage required 3GB memory; now
+  it requires about 200MB.
+
+- Use server-side cursors during packing when available, further
+  reducing the amount of memory required. See :issue:`165`.
+
 3.0a10 (2019-09-04)
 ===================